rename strings/strings.ts to strings/mod.ts
