+++ io/util.ts
@@ -2,7 +2,7 @@
 const { Buffer, mkdir, open } = Deno;
 type File = Deno.File;
 type Reader = Deno.Reader;
-import { encode } from "../strings/strings.ts";
+import { encode } from "../strings/mod.ts";
 import * as path from "../fs/path.ts";
 // `off` is the offset into `dst` where it will at which to begin writing values
 // from `src`.
