+++ io/writers.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 type Writer = Deno.Writer;
-import { decode, encode } from "../strings/strings.ts";
+import { decode, encode } from "../strings/mod.ts";
 
 /** Writer utility for buffering string chunks */
 export class StringWriter implements Writer {
