+++ strings/mod.ts
@@ -1,0 +1,5 @@
+// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
+
+export * from "./encode.ts";
+export * from "./decode.ts";
+export * from "./pad.ts";
